feat : isopen으로 계속 해서 나오던 에러 수정 완료
